+++ src/Auth/Platforms/iOS/FirebaseAuthImplementation.cs
@@ -41,7 +41,7 @@ public sealed class FirebaseAuthImplementation : DisposableBase, IFirebaseAuth
     {
         AuthErrorCode errorCode;
         if(IntPtr.Size == 8) { // 64 bits devices
-            errorCode = (AuthErrorCode) (long) ex.Error.Code;
+            errorCode = (AuthErrorCode) ex.Error.Code;
         } else { // 32 bits devices
             errorCode = (AuthErrorCode) (int) ex.Error.Code;
         }
